roblib: re-implemented hashmap based original gpio trait+backend
the new typed version is now `TypedGpio` and the original `Gpio` is back
with some adjustments. the typed version is great for humans on the
client, but on the server and in other situations, the original one is
necessary. the client now implements both,
the typed one is for human consumption, the original one is for ffi.
usage of these raw gpio calls whilst using the typed api could result in
problems but it's ok I think (?)

TODO: test the gpio backend code to make sure it works and doesn't
deadlock